Implementación de excepciones finalizada inicio de comprobación con herramienta postman

+++ src/main/java/co/edu/unbosque/controller/UsuarioRestController.java
@@ -2,6 +2,8 @@ package co.edu.unbosque.controller;
 
 import co.edu.unbosque.entity.Usuario;
 import co.edu.unbosque.service.api.UsuarioServiceAPI;
+import co.edu.unbosque.utils.exception.GeneralException;
+import co.edu.unbosque.utils.exception.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -24,34 +26,54 @@ public class UsuarioRestController {
     }
     
     @PostMapping(value = "/saveUsuario")
-    public ResponseEntity<Usuario> saveUsuario(@RequestBody Usuario usuario) {
-        return  ResponseEntity.status(HttpStatus.CREATED).body(usuarioServiceAPI.save(usuario));
+    public ResponseEntity<Usuario> saveUsuario(@RequestBody Usuario usuario) throws GeneralException {
+        try{
+           return ResponseEntity.status(HttpStatus.CREATED).body(usuarioServiceAPI.save(usuario));
+        } catch (Exception e){
+            throw new GeneralException("Error al guardar el usuario: " + e.getMessage());
+        }
     }
     
     @GetMapping(value = "/findRecord/{id}")
-    public ResponseEntity<Usuario> getUsuarioById(@PathVariable Integer id) {
-        return ResponseEntity.ok(usuarioServiceAPI.get(id));
+    public ResponseEntity<Usuario> getUsuarioById(@PathVariable Integer id) throws ResourceNotFoundException{
+        Usuario usuario = usuarioServiceAPI.get(id);
+        if(usuario == null){
+            throw new ResourceNotFoundException("usuario no encontrado con id: " + id);
+        }
+        return ResponseEntity.ok(usuario);
     }
     
     @DeleteMapping(value = "/deleteUsuario/{id}")
-    public ResponseEntity<Void> deleteUsuario(@PathVariable Integer id) {
-       usuarioServiceAPI.delete(id);
+    public ResponseEntity<Void> deleteUsuario(@PathVariable Integer id) throws ResourceNotFoundException {
+        Usuario usuario = usuarioServiceAPI.get(id);
+        if (usuario == null){
+            throw new ResourceNotFoundException("Usuario no encontrado con id: " + id);
+        }
+        usuarioServiceAPI.delete(id);
         return ResponseEntity.noContent().build();
     }
 
     @PutMapping("/updateUsuario")
-    public ResponseEntity<Usuario> updateUsuario(@RequestBody Usuario usuario){
-        Usuario user = usuarioServiceAPI.get(usuario.getIdUsuario());
-        String passwordActual = user.getPassword();
-        user.setUsername(usuario.getUsername());
-        user.setPassword(usuario.getPassword());
-        user.setNombreApellido(usuario.getNombreApellido());
-        user.setIdRol(usuario.getIdRol());
-        user.setEstado(usuario.getEstado());
-        if (!usuario.getPassword().equals(passwordActual)){
-            user.setFechaUltClave(LocalDateTime.now());
+    public ResponseEntity<Usuario> updateUsuario(@RequestBody Usuario usuario) throws ResourceNotFoundException, GeneralException{
+        try{
+            Usuario user = usuarioServiceAPI.get(usuario.getIdUsuario());
+            if(user == null){
+                throw new ResourceNotFoundException("Usuario no encontrado con id: " + usuario.getIdUsuario());
+            }
+            String passwordActual = user.getPassword();
+            user.setUsername(usuario.getUsername());
+            user.setPassword(usuario.getPassword());
+            user.setNombreApellido(usuario.getNombreApellido());
+            user.setEstado(usuario.getEstado());
+            if (!usuario.getPassword().equals(passwordActual)){
+                user.setFechaUltClave(LocalDateTime.now());
+            }
+            return ResponseEntity.ok(usuarioServiceAPI.update(user));
+        } catch (ResourceNotFoundException e){
+            throw e;
+        } catch (Exception e){
+            throw new GeneralException("Error al actualizar el usuario: " + e.getMessage());
         }
-        return ResponseEntity.ok(usuarioServiceAPI.update(user));
     }
 
 }
